refactor: 패키지 이름 수정 및 위치 이동
