+++ src/main/java/chess/console/Application.java
@@ -1,4 +1,4 @@
-package chess.game;
+package chess.console;
 
 public class Application {
