+++ src/main/java/chess/console/view/input/InputView.java
@@ -1,4 +1,4 @@
-package chess.view.input;
+package chess.console.view.input;
 
 import java.util.Arrays;
 import java.util.List;
